backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
greenlet>=3.0.0

# Authentication
PyJWT[crypto]>=2.8.0
python-jose[cryptography]>=3.3.0

# Database migrations
alembic>=1.13.0

# Testing
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Environment and configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# CORS and middleware
python-multipart>=0.0.6

# Email validation
email-validator>=2.1.0
